+++ python/dalex/dalex/_explainer/object.py
@@ -3,7 +3,7 @@ from dalex.dataset_level import ModelPerformance, VariableImportance,\
 from dalex.instance_level import BreakDown, Shap, CeterisParibus
 from dalex.wrappers import ShapWrapper
 from .checks import *
-from .utils import unpack_kwargs_lime, create_surrogate_model
+from .utils import unpack_kwargs_lime, create_surrogate_model, check_import
 
 
 class Explainer:
@@ -305,6 +305,7 @@ class Explainer:
                 random_state=random_state
             )
         elif type == 'shap_wrapper':
+            check_import('shap', msg='Install shap>=0.35.0 for SHAP explanations.')
             predict_parts_ = ShapWrapper('predict_parts')
 
         predict_parts_.fit(self, new_observation, **kwargs)
@@ -412,6 +413,7 @@ class Explainer:
         """
 
         if type == 'lime':
+            check_import('lime', msg='Install lime>=0.2.0.1 for LIME explanations.')
             from lime.lime_tabular import LimeTabularExplainer
             new_observation = check_new_observation_lime(new_observation)
 
@@ -541,7 +543,8 @@ class Explainer:
                 keep_raw_permutations=keep_raw_permutations,
             )
             model_parts_.fit(self)
-        else:
+        elif type == 'shap_wrapper':
+            check_import('shap', msg='Install shap>=0.35.0 for SHAP explanations.')
             model_parts_ = ShapWrapper('model_parts')
             if N is None:
                 N = self.data.shape[0]
@@ -728,6 +731,8 @@ class Explainer:
         https://christophm.github.io/interpretable-ml-book/global.html
         https://github.com/scikit-learn/scikit-learn
         """
+
+        check_import('scikit-learn', msg='Install scikit-learn>=0.21 for surrogate models.')
 
         types = ('tree', 'linear')
         type = check_method_type(type, types)